fix: don't configure source/execution data for projects with no dirs
This commit modifies the CodeCoverageReportTask to no longer add source
sets or execution data for projects which have no actual source
directory. This to account for projects which use intermediate sub
projects as an organisational method. These projects often have no
source directories and would cause the task to fail.